+++ js/game-10.js
@@ -1,0 +1,18 @@
+/**
+ * Created by Viacheslav on 20.11.2016.
+ */
+
+import game from './game.js';
+import nextScreen from './stats.js';
+import question from './templates/question-1.js';
+
+const data = {
+  task: 'Угадайте для каждого изображения фото или рисунок?',
+  answers: {
+    image: 'http://placehold.it/468x458',
+    photo: 'Фото',
+    paint: 'Рисунок'
+  }
+};
+
+export default game(data, question, nextScreen);
