+++ js/game-8.js
@@ -1,14 +1,15 @@
 /**
  * Created by Viacheslav on 20.11.2016.
  */
+
 import game from './game.js';
-import nextScreen from './game-4.js';
-import question from './templates/question-3.js';
+import nextScreen from './game-9.js';
+import question from './templates/question-2.js';
 
 const data = {
-  task: 'Найдите рисунок среди изображений',
+  task: 'Угадай, фото или рисунок?',
   answers: {
-    image: 'http://placehold.it/304x455',
+    image: 'http://placehold.it/705x455',
     photo: 'Фото',
     paint: 'Рисунок'
   }
