+++ js/templates/question-2.js
@@ -2,17 +2,8 @@
  * Created by Viacheslav on 27.11.2016.
  */
 
-const data = {
-  task: 'Угадай, фото или рисунок?',
-  answers: {
-    image: 'http://placehold.it/705x455',
-    photo: 'Фото',
-    paint: 'Рисунок'
-  }
-};
-
-const question = `
-      <p class="game__task">${data.task}</p>
+const question = (data) => {
+  return `<p class="game__task">${data.task}</p>
       <form class="game__content  game__content--wide">
         <div class="game__option">
           <img src="${data.answers.image}" alt="Option 1" width="705" height="455">
@@ -26,5 +17,6 @@ const question = `
           </label>
         </div>
       </form>`;
+};
 
 export default question;
